# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "LightRAG Server API",
        "description": "Providing API for LightRAG core, Web UI and Ollama Model Emulation\n\n[View ReDoc documentation](/redoc)",
        "version": "0273"
    },
    "paths": {
...
        "/documents/upload": {
            "post": {
                "tags": [
                    "documents"
                ],
                "summary": "Upload To Input Dir",
                "description": "Upload a file to the input directory and index it.\n\nThis API endpoint accepts a file through an HTTP POST request, checks if the\nuploaded file is of a supported type, saves it in the specified input directory,\nindexes it for retrieval, and returns a success status with relevant details.\n\n**File Size Limit:**\n- Configurable via `MAX_UPLOAD_SIZE` environment variable (default: 100MB)\n- Set to `None` or `0` for unlimited upload size\n- Returns HTTP 413 (Request Entity Too Large) if file exceeds limit\n\n**Duplicate Detection Behavior:**\n\nThis endpoint handles two types of duplicate scenarios differently:\n\n1. **Filename Duplicate (Synchronous Detection)**:\n   - Detected immediately before file processing\n   - Returns `status=\"duplicated\"` with the existing document's track_id\n   - Two cases:\n     - If filename exists in document storage: returns existing track_id\n     - If filename exists in file system only: returns empty track_id (\"\")\n\n2. **Content Duplicate (Asynchronous Detection)**:\n   - Detected during background processing after content extraction\n   - Returns `status=\"success\"` with a new track_id immediately\n   - The duplicate is detected later when processing the file content\n   - Use `/documents/track_status/{track_id}` to check the final result:\n     - Document will have `status=\"FAILED\"`\n     - `error_msg` contains \"Content already exists. Original doc_id: xxx\"\n     - `metadata.is_duplicate=true` with reference to original document\n     - `metadata.original_doc_id` points to the existing document\n     - `metadata.original_track_id` shows the original upload's track_id\n\n**Why Different Behavior?**\n- Filename check is fast (simple lookup), done synchronously\n- Content extraction is expensive (PDF/DOCX parsing), done asynchronously\n- This design prevents blocking the client during expensive operations\n\nArgs:\n    background_tasks: FastAPI BackgroundTasks for async processing\n    file (UploadFile): The file to be uploaded. It must have an allowed extension.\n\nReturns:\n    InsertResponse: A response object containing the upload status and a message.\n        - status=\"success\": File accepted and queued for processing\n        - status=\"duplicated\": Filename already exists (see track_id for existing document)\n\nRaises:\n    HTTPException: If the file type is not supported (400), file too large (413), or other errors occur (500).",
                "operationId": "upload_to_input_dir_documents_upload_post",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "api_key_header_value",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Api Key Header Value"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_upload_to_input_dir_documents_upload_post"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InsertResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/documents": {
            "delete": {
                "tags": [
                    "documents"
                ],
                "summary": "Clear Documents",
                "description": "Clear all documents from the RAG system.\n\nThis endpoint deletes all documents, entities, relationships, and files from the system.\nIt uses the storage drop methods to properly clean up all data and removes all files\nfrom the input directory.\n\nReturns:\n    ClearDocumentsResponse: A response object containing the status and message.\n        - status=\"success\":           All documents and files were successfully cleared.\n        - status=\"partial_success\":   Document clear job exit with some errors.\n        - status=\"busy\":              Operation could not be completed because the pipeline is busy.\n        - status=\"fail\":              All storage drop operations failed, with message\n        - message: Detailed information about the operation results, including counts\n          of deleted files and any errors encountered.\n\nRaises:\n    HTTPException: Raised when a serious error occurs during the clearing process,\n                  with status code 500 and error details in the detail field.",
                "operationId": "clear_documents_documents_delete",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "api_key_header_value",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Api Key Header Value"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClearDocumentsResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "documents"
                ],
                "summary": "Documents",
                "description": "Get the status of all documents in the system. This endpoint is deprecated; use /documents/paginated instead.\nTo prevent excessive resource consumption, a maximum of 1,000 records is returned.\n\nThis endpoint retrieves the current status of all documents, grouped by their\nprocessing status (PENDING, PROCESSING, PREPROCESSED, PROCESSED, FAILED). The results are\nlimited to 1000 total documents with fair distribution across all statuses.\n\nReturns:\n    DocsStatusesResponse: A response object containing a dictionary where keys are\n                        DocStatus values and values are lists of DocStatusResponse\n                        objects representing documents in each status category.\n                        Maximum 1000 documents total will be returned.\n\nRaises:\n    HTTPException: If an error occurs while retrieving document statuses (500).",
                "operationId": "documents_documents_get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "api_key_header_value",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Api Key Header Value"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DocsStatusesResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/documents/paginated": {
            "post": {
                "tags": [
                    "documents"
                ],
                "summary": "Get Documents Paginated",
                "description": "Get documents with pagination support.\n\nThis endpoint retrieves documents with pagination, filtering, and sorting capabilities.\nIt provides better performance for large document collections by loading only the\nrequested page of data.\n\nArgs:\n    request (DocumentsRequest): The request body containing pagination parameters\n\nReturns:\n    PaginatedDocsResponse: A response object containing:\n        - documents: List of documents for the current page\n        - pagination: Pagination information (page, total_count, etc.)\n        - status_counts: Count of documents by status for all documents\n\nRaises:\n    HTTPException: If an error occurs while retrieving documents (500).",
                "operationId": "get_documents_paginated_documents_paginated_post",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "api_key_header_value",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Api Key Header Value"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DocumentsRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginatedDocsResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Body_upload_to_input_dir_documents_upload_post": {
                "properties": {
                    "file": {
                        "type": "string",
                        "format": "binary",
                        "title": "File"
                    }
                },
                "type": "object",
                "required": [
                    "file"
                ],
                "title": "Body_upload_to_input_dir_documents_upload_post"
            },
...
            "ClearDocumentsResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": [
                            "success",
                            "partial_success",
                            "busy",
                            "fail"
                        ],
                        "title": "Status",
                        "description": "Status of the clear operation"
                    },
                    "message": {
                        "type": "string",
                        "title": "Message",
                        "description": "Message describing the operation result"
                    }
                },
                "type": "object",
                "required": [
                    "status",
                    "message"
                ],
                "title": "ClearDocumentsResponse",
                "description": "Response model for document clearing operation\n\nAttributes:\n    status: Status of the clear operation\n    message: Detailed message describing the operation result",
                "example": {
                    "message": "All documents cleared successfully. Deleted 15 files.",
                    "status": "success"
                }
            },
...
            "DocStatus": {
                "type": "string",
                "enum": [
                    "pending",
                    "processing",
                    "preprocessed",
                    "processed",
                    "failed"
                ],
                "title": "DocStatus",
                "description": "Document processing status"
            },
...
            "DocsStatusesResponse": {
                "properties": {
                    "statuses": {
                        "additionalProperties": {
                            "items": {
                                "$ref": "#/components/schemas/DocStatusResponse"
                            },
                            "type": "array"
                        },
                        "propertyNames": {
                            "$ref": "#/components/schemas/DocStatus"
                        },
                        "type": "object",
                        "title": "Statuses",
                        "description": "Dictionary mapping document status to lists of document status responses"
                    }
                },
                "type": "object",
                "title": "DocsStatusesResponse",
                "description": "Response model for document statuses\n\nAttributes:\n    statuses: Dictionary mapping document status to lists of document status responses",
                "example": {
                    "statuses": {
                        "PENDING": [
                            {
                                "content_length": 5000,
                                "content_summary": "Pending document",
                                "created_at": "2025-03-31T10:00:00",
                                "file_path": "pending_doc.pdf",
                                "id": "doc_123",
                                "status": "pending",
                                "track_id": "upload_20250331_100000_abc123",
                                "updated_at": "2025-03-31T10:00:00"
                            }
                        ],
                        "PREPROCESSED": [
                            {
                                "chunks_count": 10,
                                "content_length": 7200,
                                "content_summary": "Document pending final indexing",
                                "created_at": "2025-03-31T09:30:00",
                                "file_path": "preprocessed_doc.pdf",
                                "id": "doc_789",
                                "status": "preprocessed",
                                "track_id": "upload_20250331_093000_xyz789",
                                "updated_at": "2025-03-31T09:35:00"
                            }
                        ],
                        "PROCESSED": [
                            {
                                "chunks_count": 8,
                                "content_length": 8000,
                                "content_summary": "Processed document",
                                "created_at": "2025-03-31T09:00:00",
                                "file_path": "processed_doc.pdf",
                                "id": "doc_456",
                                "metadata": {
                                    "author": "John Doe"
                                },
                                "status": "processed",
                                "track_id": "insert_20250331_090000_def456",
                                "updated_at": "2025-03-31T09:05:00"
                            }
                        ]
                    }
                }
            },
            "DocumentsRequest": {
                "properties": {
                    "status_filter": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/DocStatus"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "description": "Filter by document status, None for all statuses"
                    },
                    "page": {
                        "type": "integer",
                        "minimum": 1.0,
                        "title": "Page",
                        "description": "Page number (1-based)",
                        "default": 1
                    },
                    "page_size": {
                        "type": "integer",
                        "maximum": 200.0,
                        "minimum": 10.0,
                        "title": "Page Size",
                        "description": "Number of documents per page (10-200)",
                        "default": 50
                    },
                    "sort_field": {
                        "type": "string",
                        "enum": [
                            "created_at",
                            "updated_at",
                            "id",
                            "file_path"
                        ],
                        "title": "Sort Field",
                        "description": "Field to sort by",
                        "default": "updated_at"
                    },
                    "sort_direction": {
                        "type": "string",
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "title": "Sort Direction",
                        "description": "Sort direction",
                        "default": "desc"
                    }
                },
                "type": "object",
                "title": "DocumentsRequest",
                "description": "Request model for paginated document queries\n\nAttributes:\n    status_filter: Filter by document status, None for all statuses\n    page: Page number (1-based)\n    page_size: Number of documents per page (10-200)\n    sort_field: Field to sort by ('created_at', 'updated_at', 'id', 'file_path')\n    sort_direction: Sort direction ('asc' or 'desc')",
                "example": {
                    "page": 1,
                    "page_size": 50,
                    "sort_direction": "desc",
                    "sort_field": "updated_at",
                    "status_filter": "PROCESSED"
                }
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "InsertResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": [
                            "success",
                            "duplicated",
                            "partial_success",
                            "failure"
                        ],
                        "title": "Status",
                        "description": "Status of the operation"
                    },
                    "message": {
                        "type": "string",
                        "title": "Message",
                        "description": "Message describing the operation result"
                    },
                    "track_id": {
                        "type": "string",
                        "title": "Track Id",
                        "description": "Tracking ID for monitoring processing status"
                    }
                },
                "type": "object",
                "required": [
                    "status",
                    "message",
                    "track_id"
                ],
                "title": "InsertResponse",
                "description": "Response model for document insertion operations\n\nAttributes:\n    status: Status of the operation (success, duplicated, partial_success, failure)\n    message: Detailed message describing the operation result\n    track_id: Tracking ID for monitoring processing status",
                "example": {
                    "message": "File 'document.pdf' uploaded successfully. Processing will continue in background.",
                    "status": "success",
                    "track_id": "upload_20250729_170612_abc123"
                }
            },
...
            "PaginatedDocsResponse": {
                "properties": {
                    "documents": {
                        "items": {
                            "$ref": "#/components/schemas/DocStatusResponse"
                        },
                        "type": "array",
                        "title": "Documents",
                        "description": "List of documents for the current page"
                    },
                    "pagination": {
                        "$ref": "#/components/schemas/PaginationInfo",
                        "description": "Pagination information"
                    },
                    "status_counts": {
                        "additionalProperties": {
                            "type": "integer"
                        },
                        "type": "object",
                        "title": "Status Counts",
                        "description": "Count of documents by status for all documents"
                    }
                },
                "type": "object",
                "required": [
                    "documents",
                    "pagination",
                    "status_counts"
                ],
                "title": "PaginatedDocsResponse",
                "description": "Response model for paginated document queries\n\nAttributes:\n    documents: List of documents for the current page\n    pagination: Pagination information\n    status_counts: Count of documents by status for all documents",
                "example": {
                    "documents": [
                        {
                            "chunks_count": 12,
                            "content_length": 15240,
                            "content_summary": "Research paper on machine learning",
                            "created_at": "2025-03-31T12:34:56",
                            "file_path": "research_paper.pdf",
                            "id": "doc_123456",
                            "metadata": {
                                "author": "John Doe",
                                "year": 2025
                            },
                            "status": "PROCESSED",
                            "track_id": "upload_20250729_170612_abc123",
                            "updated_at": "2025-03-31T12:35:30"
                        }
                    ],
                    "pagination": {
                        "has_next": true,
                        "has_prev": false,
                        "page": 1,
                        "page_size": 50,
                        "total_count": 150,
                        "total_pages": 3
                    },
                    "status_counts": {
                        "FAILED": 5,
                        "PENDING": 10,
                        "PREPROCESSED": 5,
                        "PROCESSED": 130,
                        "PROCESSING": 5
                    }
                }
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "description": "OAuth2 Password Authentication",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "login"
                    }
                }
            }
        }
    }
}